gentle git gone script
This is a non-forcible branch clean up script.
It is also an alternative to git_branchcleanup.nu

My first inclination was to put this in cool_oneliners because it is
one. But then I saw we have a few git scripts so wanted to follow them.

It raises the general question: should one liners be their own category
or can we derive an index of one liners, searched and sorted by topic?