Infraestructure initial setup.
Some groundwork to start growing Nu's standard library. As an initial exercise
we take the mantainer_time.nu script that computes the core team's table timezone
so that a given date can be converted to their local timezones.

While doing so, it sounded like the logic could be abstracted out and make it
a subcommand of `date` instead. For that reason, that particular custom command
of interested was moved to `std/date`.

Another one added were basic infraestructure custom commands for the test framework,
subject to change (naming, and more things). It's in `std/playground` and it's used
in `tests` directory.

Since this repository contains scripts, we keep the `language` directory to maintain
and grow the nu std library and grow it as we see custom command candidates.

For the majority of the scripts that have nothing to do with the standard language but
they do exercise Nu in practice (and some use custom commands, ideally, from the very, hopefully,
future nu standard library) I have expectations to make them `runnable` as well and for that reason
we introduce here an `examples` directory.

To run the first one, you can do:

```
nu examples/date_in_local_timezones.nu
```