fix pass-completion as `let` cannot be used at module level (#544)
* fix pass-completion as `let` cannot be used at module level

`let` cannont be used at module level. #8248 proposes to add `const` for modules, but in the meantime, a function is the best way to make this script useable as a module.

* remove calls to system commands for better portability

`^realpath` and `^find` are replaced with `path expand` and `ls` in an effort not to rely on system commands. Literal "/" are also replaced with native path operations.

---------

Co-authored-by: Benoît Sierro <benoit.sierro@bluewin.ch>