Adding http oneliner and changing reference to renamed Discord channel. (#731)
The [cookbook intro page](https://www.nushell.sh/cookbook/) gives an
example of a oneliner that I recently changed to imply is one of the
oneliners from this repo (nushell/nushell.github.io#1205), however that
oneliner isn't actually in this repo. This adds that oneliner to the
repo.

I also replaced the mention of the \#cool-oneliners channel that was
renamed to \#cool-scripts to just mention the general Nushell community.