Feature/add git grep completions (#898)
Added completions for most of the flags for `git grep`. 

A few flags have been omitted for this PR due to Git expecting arguments
for those be passed without whitespaces. The arguments are optional, but
the flags would not work when passed arguments (see `--open-with-pager`
and `--color` flags in [this
commit](https://github.com/nushell/nu_scripts/commit/ffa9ceab737a575397f625044143c9f26f49ac97)
for how I tried to do it).

Ex. 

`--open-with-pager(-O)` expects the optional argument to be passed as
either `--open-with-pager=<pager>` or `-O<pager>`. Including it in the
completions list as I had implemented it breaks the flag when using the
optional parameter.

Feedback appreciated for how to handle that. I'm still new to the Nu
language, so I might have missed something obvious that could fix that.